Add project reorder functionality
This adds a bunch of code in the projects dom manager, main script, and
database script to facilitate this. I also realized that it was not
important to rearange tasks nilly willy, so I did not implement it
for the tasks. If I decide that it is needed, I can add it later.

resolves: #1