Change way id is excluded from updating
Before, I would simply delete the id from the object that was
used for updating. If the object was used for something after
the task or project was updated, the id would be missing since
it was all linked together. Now, I just ignore it while keeping
it in the object.

This was mainly a problem for the task class, but I also moved
the changes to the project class anyways.