Substatial, comlex change (see notes)
This did a bunch of things that I should have broken into different
commits.

This gets the task info modal up and going. All functionality work
for it now. It is styled generally appropriately too.

This adds a bunch of helper methods to the different DOM manager
modules and other modules too.

Lastly, this changed some of the names of functions to be more uniform.
Insead of having "removed", "deleted", "added", "updated", and others,
I tried to pair down to the basic terminology.

This also included a date package that is used to help format dates.